Addressing comments - 1
