+++ error/s2n_errno.h
@@ -262,6 +262,7 @@ typedef enum {
     S2N_ERR_ASYNC_ALREADY_APPLIED,
     S2N_ERR_UNSUPPORTED_WITH_QUIC,
     S2N_ERR_DUPLICATE_PSK_IDENTITIES,
+    S2N_ERR_VALID_PSK_IDENTITY_NOT_FOUND,
     S2N_ERR_T_USAGE_END,
 } s2n_error;
