+++ tests/unit/s2n_client_psk_extension_test.c
@@ -26,9 +26,6 @@
 #define TEST_BYTES 0x01, 0xFF, 0x23
 #define TEST_BYTES_SIZE 0x00, 0x03
 
-#define TEST_BYTES_2 0x0A, 0x0B, 0x0C
-#define TEST_BYTES_SIZE_2 0x00, 0x03
-
 struct s2n_psk_test_case {
     s2n_hmac_algorithm hmac_alg;
     uint8_t hash_size;
@@ -36,47 +33,11 @@ struct s2n_psk_test_case {
     size_t identity_size;
 };
 
-static S2N_RESULT verify_chosen_psk_equals(struct s2n_psk *chosen_psk, struct s2n_psk *match_psk)
-{
-    ENSURE_REF(chosen_psk);
-    ENSURE_REF(match_psk);
-
-    ENSURE_EQ(memcmp(chosen_psk->identity.data, match_psk->identity.data, match_psk->identity.size), 0);
-    ENSURE_EQ(chosen_psk->identity.size, match_psk->identity.size);
-    ENSURE_EQ(memcmp(chosen_psk->secret.data, match_psk->secret.data, match_psk->secret.size), 0);
-    ENSURE_EQ(chosen_psk->secret.size, match_psk->secret.size);
-    ENSURE_EQ(chosen_psk->hmac_alg, match_psk->hmac_alg);
-    ENSURE_EQ(chosen_psk->type, S2N_PSK_TYPE_EXTERNAL);
-    ENSURE_EQ(chosen_psk->obfuscated_ticket_age, 0);
-    ENSURE_EQ(chosen_psk->early_secret.data, NULL);
-    ENSURE_EQ(chosen_psk->early_secret.size, 0);
-
-    return S2N_RESULT_OK;
-}
-
-static S2N_RESULT verify_external_chosen_psk_equals(struct s2n_external_psk *ex_chosen_psk, struct s2n_psk *match_psk)
-{
-    ENSURE_REF(ex_chosen_psk);
-    ENSURE_REF(match_psk);
-
-    struct s2n_external_psk test_ex_psk = { 0 };
-    GUARD_AS_RESULT(s2n_external_psk_set_hmac(&test_ex_psk, match_psk->hmac_alg));
-
-    ENSURE_EQ(memcmp(ex_chosen_psk->identity, match_psk->identity.data, match_psk->identity.size), 0);
-    ENSURE_EQ(ex_chosen_psk->identity_length, match_psk->identity.size);
-    ENSURE_EQ(memcmp(ex_chosen_psk->secret, match_psk->secret.data, match_psk->secret.size), 0);
-    ENSURE_EQ(ex_chosen_psk->secret_length, match_psk->secret.size);
-    ENSURE_EQ(ex_chosen_psk->hmac, test_ex_psk.hmac);
-
-    return S2N_RESULT_OK;
-}
-
 int main(int argc, char **argv)
 {
     BEGIN_TEST();
 
     uint8_t test_bytes_data[] = { TEST_BYTES };
-    uint8_t test_bytes_data_2[] = { TEST_BYTES_2 };
     uint8_t test_identity[] = "test identity";
     uint8_t test_identity_2[] = "another identity";
     uint8_t test_secret[] = "test secret";
@@ -344,7 +305,6 @@ int main(int argc, char **argv)
             struct s2n_stuffer wire_identities_in = { 0 };
             EXPECT_SUCCESS(s2n_stuffer_alloc(&wire_identities_in, sizeof(single_wire_identity)));
             EXPECT_SUCCESS(s2n_stuffer_write_bytes(&wire_identities_in, single_wire_identity, sizeof(single_wire_identity)));
-
             uint16_t identity_count = 0;
 
             EXPECT_OK(s2n_count_psk_identities(&wire_identities_in, &identity_count));
@@ -358,164 +318,12 @@ int main(int argc, char **argv)
             struct s2n_stuffer wire_identities_in = { 0 };
             EXPECT_SUCCESS(s2n_stuffer_alloc(&wire_identities_in, sizeof(wire_identites)));
             EXPECT_SUCCESS(s2n_stuffer_write_bytes(&wire_identities_in, wire_identites, sizeof(wire_identites)));
-
             uint16_t identity_count = 0;
 
             EXPECT_OK(s2n_count_psk_identities(&wire_identities_in, &identity_count));
             EXPECT_EQUAL(identity_count, 5);
 
             EXPECT_SUCCESS(s2n_stuffer_free(&wire_identities_in));
-        }
-    }
-
-    /* Test: s2n_match_psk_identity */
-    {
-        struct s2n_psk_parameters params = { 0 };
-        EXPECT_OK(s2n_psk_parameters_init(&params));
-
-        struct s2n_array *known_psks = &params.psk_list;
-
-        struct s2n_blob wire_identity = { 0 };
-        EXPECT_SUCCESS(s2n_blob_init(&wire_identity, test_identity, sizeof(test_identity)));
-
-        /* Test: No known PSKs */
-        {
-            struct s2n_psk *match = NULL;
-            EXPECT_OK(s2n_match_psk_identity(known_psks, &wire_identity, &match));
-            EXPECT_NULL(match);
-        }
-
-        /* Test: No match exists */
-        {
-            struct s2n_psk *different_identity = NULL;
-            EXPECT_OK(s2n_array_pushback(known_psks, (void**) &different_identity));
-            EXPECT_SUCCESS(s2n_psk_init(different_identity, S2N_PSK_TYPE_EXTERNAL));
-            EXPECT_SUCCESS(s2n_psk_new_identity(different_identity, test_identity_2, sizeof(test_identity_2)));
-
-            struct s2n_psk *match = NULL;
-            EXPECT_OK(s2n_match_psk_identity(known_psks, &wire_identity, &match));
-            EXPECT_NULL(match);
-        }
-
-        struct s2n_psk *expected_match;
-
-        /* Test: Match exists */
-        {
-            EXPECT_OK(s2n_array_pushback(known_psks, (void**) &expected_match));
-            EXPECT_SUCCESS(s2n_psk_init(expected_match, S2N_PSK_TYPE_EXTERNAL));
-            EXPECT_SUCCESS(s2n_psk_new_identity(expected_match, test_identity, sizeof(test_identity)));
-
-            struct s2n_psk *match = NULL;
-            EXPECT_OK(s2n_match_psk_identity(known_psks, &wire_identity, &match));
-            EXPECT_EQUAL(match, expected_match);
-        }
-
-        /* Test: Multiple matches exist */
-        {
-            struct s2n_psk *another_match = NULL;
-            EXPECT_OK(s2n_array_pushback(known_psks, (void**) &another_match));
-            EXPECT_SUCCESS(s2n_psk_init(another_match, S2N_PSK_TYPE_EXTERNAL));
-            EXPECT_SUCCESS(s2n_psk_new_identity(another_match, test_identity, sizeof(test_identity)));
-
-            struct s2n_psk *match = NULL;
-            EXPECT_OK(s2n_match_psk_identity(known_psks, &wire_identity, &match));
-            EXPECT_EQUAL(match, expected_match);
-        }
-
-        EXPECT_OK(s2n_psk_parameters_wipe(&params));
-    }
-
-    /* Test: s2n_choose_psk */
-    { 
-        /* Safety checks */
-        {
-            struct s2n_psk_identity psk_identities[] = { 0 };
-            struct s2n_connection *conn;
-            EXPECT_NOT_NULL(conn = s2n_connection_new(S2N_SERVER));
-
-            struct s2n_chosen_psk chosen_psk = { 0 };
-            size_t psk_identities_length = 0; 
-
-            EXPECT_FAILURE_WITH_ERRNO(s2n_choose_psk(NULL, psk_identities, psk_identities_length, &chosen_psk),
-                                      S2N_ERR_NULL);
-
-            EXPECT_FAILURE_WITH_ERRNO(s2n_choose_psk(conn, NULL, psk_identities_length, &chosen_psk), S2N_ERR_NULL);
-
-            EXPECT_FAILURE_WITH_ERRNO(s2n_choose_psk(conn, psk_identities, psk_identities_length, NULL), S2N_ERR_NULL);
-
-            EXPECT_SUCCESS(s2n_connection_free(conn));
-        }
-
-        /* Choose psk from a list without a match */
-        {
-            struct s2n_connection *conn;
-            EXPECT_NOT_NULL(conn = s2n_connection_new(S2N_SERVER));
-
-            struct s2n_psk_identity psk_identities[] = { 0 };
-            psk_identities[0].data = test_bytes_data;
-            psk_identities[0].length = sizeof(test_bytes_data);
-
-            struct s2n_chosen_psk chosen_psk = { 0 };
-            EXPECT_SUCCESS(s2n_choose_psk(conn, psk_identities, 1, &chosen_psk));
-            EXPECT_NULL(conn->psk_params.chosen_psk);
-
-            EXPECT_SUCCESS(s2n_connection_free(conn));
-        }
-
-        /* Choose psk from a list with an immediate match */
-        {
-            struct s2n_connection *conn;
-            EXPECT_NOT_NULL(conn = s2n_connection_new(S2N_SERVER));
-
-            struct s2n_psk_identity psk_identities[] = { 0 };
-            psk_identities[0].data = test_bytes_data;
-            psk_identities[0].length = sizeof(test_bytes_data);
-        
-            struct s2n_psk *match_psk = NULL;
-            EXPECT_OK(s2n_array_pushback(&conn->psk_params.psk_list, ( void ** )&match_psk));
-            EXPECT_SUCCESS(s2n_psk_init(match_psk, S2N_PSK_TYPE_EXTERNAL));
-            EXPECT_SUCCESS(s2n_psk_new_identity(match_psk, test_bytes_data, sizeof(test_bytes_data)));
-
-            struct s2n_chosen_psk chosen_psk = { 0 };
-            uint8_t external_identity[3] = { 0 };
-            uint8_t external_secret[3] = { 0 };
-            chosen_psk.external_psk.identity = external_identity;
-            chosen_psk.external_psk.secret = external_secret;
-
-            EXPECT_SUCCESS(s2n_choose_psk(conn, psk_identities, 1, &chosen_psk));
-            EXPECT_EQUAL(chosen_psk.wire_index, 0);
-            EXPECT_OK(verify_external_chosen_psk_equals(&chosen_psk.external_psk, match_psk));
-
-            EXPECT_SUCCESS(s2n_connection_free(conn));
-        }
-
-        /* Choose psk with a match later in the list */
-        {
-            struct s2n_connection *conn;
-            EXPECT_NOT_NULL(conn = s2n_connection_new(S2N_SERVER));
-
-            struct s2n_psk_identity psk_identities[2] = { 0 };
-            psk_identities[0].data = test_bytes_data;
-            psk_identities[0].length = sizeof(test_bytes_data);
-            psk_identities[1].data = test_bytes_data_2;
-            psk_identities[1].length = sizeof(test_bytes_data_2);
-
-            struct s2n_psk *match_psk = NULL;
-            EXPECT_OK(s2n_array_pushback(&conn->psk_params.psk_list, ( void ** )&match_psk));
-            EXPECT_SUCCESS(s2n_psk_init(match_psk, S2N_PSK_TYPE_EXTERNAL));
-            EXPECT_SUCCESS(s2n_psk_new_identity(match_psk, test_bytes_data_2, sizeof(test_bytes_data_2)));
-
-            struct s2n_chosen_psk chosen_psk = { 0 };
-            uint8_t external_identity[3] = { 0 };
-            uint8_t external_secret[3] = { 0 };
-            chosen_psk.external_psk.identity = external_identity;
-            chosen_psk.external_psk.secret = external_secret;
-
-            EXPECT_SUCCESS(s2n_choose_psk(conn, psk_identities, 2, &chosen_psk));
-            EXPECT_EQUAL(chosen_psk.wire_index, 1);
-            EXPECT_OK(verify_external_chosen_psk_equals(&chosen_psk.external_psk, match_psk));
-
-            EXPECT_SUCCESS(s2n_connection_free(conn));
         }
     }
 
@@ -544,7 +352,8 @@ int main(int argc, char **argv)
             EXPECT_SUCCESS(s2n_stuffer_alloc(&wire_identities_in, sizeof(wire_identites)));
             EXPECT_SUCCESS(s2n_stuffer_write_bytes(&wire_identities_in, wire_identites, sizeof(wire_identites)));
 
-            EXPECT_OK(s2n_client_psk_recv_identity_list(conn, &wire_identities_in));
+            EXPECT_ERROR_WITH_ERRNO(s2n_client_psk_recv_identity_list(conn, &wire_identities_in),
+                                      S2N_ERR_VALID_PSK_IDENTITY_NOT_FOUND);
             EXPECT_NULL(conn->psk_params.chosen_psk);
             EXPECT_EQUAL(conn->psk_params.chosen_psk_wire_index, 0);
 
@@ -568,7 +377,7 @@ int main(int argc, char **argv)
 
             EXPECT_OK(s2n_client_psk_recv_identity_list(conn, &wire_identities_in));
             EXPECT_EQUAL(conn->psk_params.chosen_psk_wire_index, 0);
-            EXPECT_OK(verify_chosen_psk_equals(conn->psk_params.chosen_psk, match_psk));
+            EXPECT_EQUAL(conn->psk_params.chosen_psk, match_psk);
 
             EXPECT_SUCCESS(s2n_connection_free(conn));
             EXPECT_SUCCESS(s2n_stuffer_free(&wire_identities_in));
@@ -590,7 +399,7 @@ int main(int argc, char **argv)
 
             EXPECT_OK(s2n_client_psk_recv_identity_list(conn, &wire_identities_in));
             EXPECT_EQUAL(conn->psk_params.chosen_psk_wire_index, wire_identities_match_index);
-            EXPECT_OK(verify_chosen_psk_equals(conn->psk_params.chosen_psk, match_psk));
+            EXPECT_EQUAL(conn->psk_params.chosen_psk, match_psk);
 
             EXPECT_SUCCESS(s2n_connection_free(conn));
             EXPECT_SUCCESS(s2n_stuffer_free(&wire_identities_in));
@@ -601,8 +410,8 @@ int main(int argc, char **argv)
             struct s2n_connection *conn;
             EXPECT_NOT_NULL(conn = s2n_connection_new(S2N_SERVER));
 
-            EXPECT_SUCCESS(s2n_config_choose_psk_cb(conn, s2n_choose_psk));
-            EXPECT_EQUAL(conn->config->psk_selection_cb, s2n_choose_psk);
+            EXPECT_SUCCESS(s2n_config_set_psk_selection_callback(conn, s2n_select_psk_identity));
+            EXPECT_EQUAL(conn->config->psk_selection_cb, s2n_select_psk_identity);
 
             struct s2n_stuffer wire_identities_in = { 0 };
             EXPECT_SUCCESS(s2n_stuffer_alloc(&wire_identities_in, sizeof(wire_identites)));
@@ -615,7 +424,7 @@ int main(int argc, char **argv)
 
             EXPECT_OK(s2n_client_psk_recv_identity_list(conn, &wire_identities_in));
             EXPECT_EQUAL(conn->psk_params.chosen_psk_wire_index, wire_identities_match_index);
-            EXPECT_OK(verify_chosen_psk_equals(conn->psk_params.chosen_psk, match_psk));
+            EXPECT_EQUAL(conn->psk_params.chosen_psk, match_psk);
 
             EXPECT_SUCCESS(s2n_connection_free(conn));
             EXPECT_SUCCESS(s2n_stuffer_free(&wire_identities_in));
@@ -969,7 +778,7 @@ int main(int argc, char **argv)
 
             EXPECT_SUCCESS(s2n_client_psk_recv(server_conn, server_in));
             EXPECT_EQUAL(server_conn->psk_params.chosen_psk_wire_index, 0);
-            EXPECT_OK(verify_chosen_psk_equals(server_conn->psk_params.chosen_psk, shared_psk));
+            EXPECT_EQUAL(server_conn->psk_params.chosen_psk, shared_psk);
 
             EXPECT_SUCCESS(s2n_connection_free(server_conn));
             EXPECT_SUCCESS(s2n_connection_free(client_conn));
@@ -1027,7 +836,7 @@ int main(int argc, char **argv)
 
         /* Verify shared PSK chosen */
         EXPECT_EQUAL(server_conn->psk_params.chosen_psk_wire_index, 1);
-        EXPECT_OK(verify_chosen_psk_equals(server_conn->psk_params.chosen_psk, shared_psk));
+        EXPECT_EQUAL(server_conn->psk_params.chosen_psk, shared_psk);
         EXPECT_EQUAL(shared_psk->secret.size, sizeof(test_secret));
         EXPECT_BYTEARRAY_EQUAL(shared_psk->secret.data, test_secret, sizeof(test_secret));
